Separating logic between Windows and Linux for the controller
