+++ src/System.Device.Gpio/System.Device.Gpio.csproj
@@ -14,7 +14,13 @@
     <None Remove="winmd\**" />
     <PackageReference Include="System.Threading.Tasks.Extensions" Version="4.5.1" />
   </ItemGroup>
+  <ItemGroup Condition="'$(TargetsLinux)' == 'true'">
+    <!--Excluding Windows implementations-->
+    <Compile Remove="**\*.Windows.cs" />
+  </ItemGroup>
   <ItemGroup Condition="'$(TargetsWindows)' == 'true'">
+    <!--Excluding Linux implementations-->
+    <Compile Remove="**\*.Linux.cs" />
     <Reference Include="Windows.Devices">
       <HintPath>winmd\Windows.Devices.winmd</HintPath>
       <IsWinMDFile>true</IsWinMDFile>
